[DEVOPS-2841] Create a simple plugin for remarkable markdown parser
This is a POC to learn how to create a new plugin/extension for
remarkable markdown parser. In this POC, we created a plugin that
can detect a new syntax and produce an expected html.

Necessary unit test has also been included. Next we need to finalize
the markdown syntax we want to use and start implementing it based
on oEmbed specs.